+++ libmpdemux/demux_vqf.c
@@ -165,7 +165,8 @@ static demuxer_t* demux_open_vqf(demuxer_t* demuxer) {
     {
     demuxer->movi_start=stream_tell(s);
     demuxer->movi_end=demuxer->movi_start+chunk_size;
-    mp_msg(MSGT_DEMUX, MSGL_V, "Found data at %"PRIX64" size %"PRIu64"\n",demuxer->movi_start,demuxer->movi_end);
+    mp_msg(MSGT_DEMUX, MSGL_V, "Found data at %"PRIX64" size %"PRIu64"\n",
+           (uint64_t)demuxer->movi_start,(uint64_t)demuxer->movi_end);
     /* Done! play it */
     break;
     }
